Genericize executor test helpers
